requirements.txt
python-dotenv==0.15.0  # Use the correct version number here
streamlit==1.16.0
pandas==1.5.3
torch==2.1.0
transformers==4.47.1
pandas==2.1.4
numpy==1.26.4